Extract mobile nav logic into useMobileNav hook
Reduces Header component complexity by moving state, body
class side-effects, and keyboard focus trap into a reusable
hook. Also fixes a bug where hidden desktop menu items were
included in the mobile focus trap.
